requirements.txt
Django
djangorestframework
drf_spectacular
Faker
requests
python-dotenv
psycopg2-binary
django-cors-headers
sentry-sdk
boto3
pillow
environs
cryptomus
celery
django-celery-beat
django-celery-results
redis
django-storages
tenacity
loguru
asyncpg
httpx
stripe